fix(sidebar): Fixed a bug that some subtitles were not displayed in sidebar
Fixed a bug that some subtitles were not displayed in the subtitle sidebar.

The virtual scroll is enabled in the ListBox, but due to this, when referencing the parent DataContext using RelativeSource in the element in the ListBoxItem, it could not be resolved and became UnsetValue, which was passed to Converter.

This problem was resolved by referencing the ListBox directly by ElementName instead of RelativeSource.