fix(seek): prevent external subtitles from blinking during seek [Fixes #41]
Stop running threads in Demuxer and Decoder respectively for external subtitles, and use subtitle data loaded in memory.
Handling threads in text subtitles is expensive and I want to keep the logic simple, so do not use threads.

There is a concern that bitmap external subtitles may increase memory usage, but this is a rare case and is acceptable.