fix(subtitle): fix an issue where the first bitmap sub could not be displayed correctly
There was a problem that the first bitmap subtitle could not be displayed in the correct position
and size because it could not get the width and height of the video, due to the fact
that analyseduration and probesize are no longer increased.
This problem has been solved by caching the video size and using it.

https://github.com/SuRGeoNix/Flyleaf/issues/502